+++ src/main/java/org/bytedeco/javacv/FFmpegLogCallback.java
@@ -23,6 +23,7 @@
 package org.bytedeco.javacv;
 
 import org.bytedeco.javacpp.BytePointer;
+import org.bytedeco.javacpp.PointerScope;
 import org.bytedeco.javacpp.tools.Logger;
 
 import static org.bytedeco.javacpp.avutil.*;
@@ -39,6 +40,12 @@ public class FFmpegLogCallback extends LogCallback {
     private static final Logger logger = Logger.create(FFmpegLogCallback.class);
 
     static final FFmpegLogCallback instance = new FFmpegLogCallback();
+    static {
+        PointerScope s = PointerScope.getInnerScope();
+        if (s != null) {
+            s.detach(instance);
+        }
+    }
 
     /** Returns an instance that can be used with {@link #setLogCallback(LogCallback)}. */
     public static FFmpegLogCallback getInstance() {
@@ -62,6 +69,7 @@ public class FFmpegLogCallback extends LogCallback {
                 break;
             case AV_LOG_INFO:
                 logger.info(msg.getString());
+                break;
             case AV_LOG_VERBOSE:
             case AV_LOG_DEBUG:
             case AV_LOG_TRACE: