+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Keep globally shared callback objects for FFmpeg out of `PointerScope` ([issue #911](https://github.com/bytedeco/javacv/issues/911))
  * Upgrade dependencies for OpenCV 3.4.3, FFmpeg 4.0.2, and Tesseract 4.0.0-beta.4
  * Update the `Demo` class to use the C++ API of OpenCV ([issue #1042](https://github.com/bytedeco/javacv/issues/1042))
  * Add new `DeepLearningFaceDetection` sample ([pull #1041](https://github.com/bytedeco/javacv/pull/1041))